agent/pyproject.toml
[project]
name = "kwami-lk-agent"
version = "0.1.0"
description = "LiveKit Cloud agent for Kwami AI"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.3.0",
    "livekit-plugins-openai>=1.3.0",
    "livekit-plugins-deepgram>=1.3.0",
    "livekit-plugins-cartesia>=1.3.0",
    "livekit-plugins-elevenlabs>=1.3.0",
    "livekit-plugins-silero>=1.3.0",
    "python-dotenv>=1.0.0",
    "zep-cloud>=3.0.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]
elevenlabs = ["livekit-plugins-elevenlabs>=1.3.0"]
google = ["livekit-plugins-google>=1.3.0"]
anthropic = ["livekit-plugins-anthropic>=1.3.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
# Force rebuild v2
